# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cats/{id}": {
      "get": {
        "operationId": "CatsController_findOne",
        "summary": "IDで猫データを取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "猫データ"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CatsController_update",
        "summary": "猫データを更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "猫データが正常に更新されました"
          },
          "400": {
            "description": "無効なデータです"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CatsController_remove",
        "summary": "猫データを削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "猫データのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "猫データが正常に削除されました"
          },
          "404": {
            "description": "猫データが見つかりません"
          }
        },
        "tags": [
          "Cats"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/tags": {
      "get": {
        "operationId": "TagsController_findAll",
        "summary": "タグ一覧の取得",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "非アクティブなタグを含めるか",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "対象スコープ",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "TagsController_create",
        "summary": "タグの作成",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/reorder": {
      "patch": {
        "operationId": "TagsController_reorder",
        "summary": "タグの並び替え",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTagsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/{id}": {
      "patch": {
        "operationId": "TagsController_update",
        "summary": "タグの更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TagsController_remove",
        "summary": "タグの削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/cats/{id}/tags": {
      "post": {
        "operationId": "TagsController_assign",
        "summary": "猫にタグを付与",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "付与成功（重複時もOK）"
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/cats/{id}/tags/{tagId}": {
      "delete": {
        "operationId": "TagsController_unassign",
        "summary": "猫からタグを剥奪",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/categories": {
      "get": {
        "operationId": "TagCategoriesController_findAll",
        "summary": "タグカテゴリ一覧の取得",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "非アクティブカテゴリを含める",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "対象スコープ",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ]
      },
      "post": {
        "operationId": "TagCategoriesController_create",
        "summary": "タグカテゴリの作成",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/categories/reorder": {
      "patch": {
        "operationId": "TagCategoriesController_reorder",
        "summary": "タグカテゴリの並び替え",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderTagCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tags/categories/{id}": {
      "patch": {
        "operationId": "TagCategoriesController_update",
        "summary": "タグカテゴリの更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TagCategoriesController_remove",
        "summary": "タグカテゴリの削除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Tags"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Cat Management System API",
    "description": "API for managing cat breeding and care records",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateCatDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "タグ名",
            "example": "Indoor"
          },
          "categoryId": {
            "type": "string",
            "description": "カテゴリID",
            "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          },
          "color": {
            "type": "string",
            "description": "カラーコード",
            "example": "#3B82F6"
          },
          "description": {
            "type": "string",
            "description": "説明",
            "example": "室内飼いタグ"
          },
          "allowsManual": {
            "type": "boolean",
            "description": "手動操作で利用可能か",
            "example": true
          },
          "allowsAutomation": {
            "type": "boolean",
            "description": "自動ルールで利用可能か",
            "example": true
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順",
            "example": 10
          },
          "metadata": {
            "type": "object",
            "description": "任意のメタデータ"
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブかどうか",
            "example": true
          }
        },
        "required": [
          "name",
          "categoryId"
        ]
      },
      "TagOrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "タグID",
            "format": "uuid"
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順",
            "example": 12
          },
          "categoryId": {
            "type": "string",
            "description": "所属カテゴリID",
            "format": "uuid"
          }
        },
        "required": [
          "id",
          "displayOrder"
        ]
      },
      "ReorderTagsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagOrderItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "タグ名",
            "example": "Indoor"
          },
          "categoryId": {
            "type": "string",
            "description": "カテゴリID",
            "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          },
          "color": {
            "type": "string",
            "description": "カラーコード",
            "example": "#3B82F6"
          },
          "description": {
            "type": "string",
            "description": "説明",
            "example": "室内飼いタグ"
          },
          "allowsManual": {
            "type": "boolean",
            "description": "手動操作で利用可能か",
            "example": true
          },
          "allowsAutomation": {
            "type": "boolean",
            "description": "自動ルールで利用可能か",
            "example": true
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順",
            "example": 10
          },
          "metadata": {
            "type": "object",
            "description": "任意のメタデータ"
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブかどうか",
            "example": true
          }
        }
      },
      "AssignTagDto": {
        "type": "object",
        "properties": {
          "tagId": {
            "type": "string",
            "description": "タグID",
            "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          }
        },
        "required": [
          "tagId"
        ]
      },
      "CreateTagCategoryDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "ユニークキー (未指定時は名前から生成)",
            "example": "cats_status"
          },
          "name": {
            "type": "string",
            "description": "カテゴリ名",
            "example": "猫ステータス"
          },
          "description": {
            "type": "string",
            "description": "カテゴリの説明"
          },
          "color": {
            "type": "string",
            "description": "カテゴリの代表カラー",
            "example": "#6366F1"
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順"
          },
          "scopes": {
            "description": "利用するスコープ一覧",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブかどうか",
            "example": true
          }
        },
        "required": [
          "name"
        ]
      },
      "TagCategoryOrderItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "カテゴリID",
            "format": "uuid"
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順",
            "example": 10
          }
        },
        "required": [
          "id",
          "displayOrder"
        ]
      },
      "ReorderTagCategoriesDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagCategoryOrderItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateTagCategoryDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "ユニークキー (未指定時は名前から生成)",
            "example": "cats_status"
          },
          "name": {
            "type": "string",
            "description": "カテゴリ名",
            "example": "猫ステータス"
          },
          "description": {
            "type": "string",
            "description": "カテゴリの説明"
          },
          "color": {
            "type": "string",
            "description": "カテゴリの代表カラー",
            "example": "#6366F1"
          },
          "displayOrder": {
            "type": "number",
            "description": "表示順"
          },
          "scopes": {
            "description": "利用するスコープ一覧",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブかどうか",
            "example": true
          }
        }
      }
    }
  }
}